water-detect-backend/requirements.txt
Django==5.1.5
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==1.7.1
PyYAML==6.0.2
Pygments==2.19.1
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
django-filter==25.1
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipython==8.12.3
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.2
moviepy==2.1.2
mpmath==1.3.0
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.1
onnx==1.17.0
onnxruntime==1.20.1
onnxslim==0.1.47
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pika==1.3.2
pika-stubs==0.1.3
pillow==10.4.0
pillow==11.1.0
pip==23.2.1
pipreqs==0.5.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyparsing==3.2.1
pyreadline3==3.5.4
pyrocketmq==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==309
pyzmq==26.3.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.23.1
scipy==1.15.1
seaborn==0.13.2
setuptools==68.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tinycss2==1.4.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.63
ultralytics-thop==2.0.14
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.41.2
yarg==0.1.9
